requirements.txt
torch>=1.9.0
networkx>=2.6.0
scikit-learn>=1.0.0
pandas>=1.3.0
numpy>=1.20.0
tqdm>=4.60.0
pyyaml>=6.0
matplotlib>=3.5.0
seaborn>=0.11.0